+++ open-inference-runtime/src/lib.rs
@@ -5,7 +5,6 @@ pub mod models;
 
 pub use client::{TritonClient,TensorData};
 pub use models::ModelExtractor;
-pub use models::verify_model_blob;
 
 // #[cfg(test)]
 // mod tests;
